Add tips about adding duo_auth_required to /admin
Will address #24. Had to take the approach of giving "various options," because there isn't a single best solution for all Django app maintainers. This is partly due to constraints in Django. Did my best to help explain the options while displaying a preference to a certain option (enable 2fa 'everywhere').